requirements.txt
numpy>=1.20.0
pandas>=1.3.3
scikit-learn>=0.24.2
flask>=2.0.1
flask-cors>=3.0.10
nltk>=3.7
plotly>=5.3.1
google-generativeai>=0.3.0
hdbcli>=2.17.14
python-dotenv>=0.19.0
requests>=2.26.0
